Create WMI_Hunter.ps1
Queries a windows device for WMI entries. 

Writes to event log (you will have to set which log it writes to specifically). If you can create a way to automate this script running regularly, you will be able to create alerts for whenever a device is found with an entry like this.